+++ projects/i-novation/ng-rapidforms/material/src/public_api.ts
@@ -4,3 +4,6 @@ export * from './lib/rapidforms-material.module';
 // Material
 export * from './lib/templates/dynamicform/dynamicform-material.component';
 export * from './lib/templates/dynamicformelement/dynamicformelement-material.component';
+
+// Material models
+export * from './lib/models/elements/MaterialNumberElement';